.docker-pyannote-server/requirements.txt
# backend

pydantic==2.9.1
requests==2.32.3

# ai

pyannote.audio==3.3.2
pydub==0.25.1

transformers==4.44.2
accelerate==0.34.2
bitsandbytes==0.43.3
sentencepiece==0.2.0
sacremoses==0.1.1

# api

fastapi[all]==0.114.2
pydantic==2.9.1
uvicorn==0.30.6
